# openapi.yaml
openapi: 3.0.3
info:
  title: API de Charadas
  description: API para gerenciamento de charadas com autenticação via token.
  version: "1.0"
  contact:
    name: Isa Jucá

servers:
  - url: https://api-charadas-five.vercel.app/

paths:

  /:
    get:
      summary: Rota inicial
      responses:
        '200':
          description: Informações da API
          content:
            application/json:
              schema:
                type: object
                properties:
                  api:
                    type: string
                  version:
                    type: string
                  Author:
                    type: string

  /login:
    post:
      summary: Realizar login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usuario
                - senha
              properties:
                usuario:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '400':
          description: Dados inválidos

  /charadas:
    get:
      summary: Listar todas as charadas
      responses:
        '200':
          description: Lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'

    post:
      summary: Criar nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '201':
          description: Charada criada com sucesso
        '400':
          description: Erro na criação

  /charadas/aleatorias:
    get:
      summary: Retornar uma charada aleatória
      responses:
        '200':
          description: Charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    get:
      summary: Buscar charada por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada

    put:
      summary: Atualizar totalmente uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
        '404':
          description: Charada não encontrada

    patch:
      summary: Atualizar parcialmente uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
        '404':
          description: Charada não encontrada

    delete:
      summary: Deletar uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Excluído com sucesso
        '404':
          description: Charada não encontrada

components:
  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
        pergunta:
          type: string
        resposta:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT